requirements.txt
anthropic
click
rich
tabulate
python-dotenv
